requirements.txt
streamlit
requests
pyyaml
pandas
numpy
